refactor summarizable query handling (#4448)
