+++ app/helpers/avo/summary_chart_helper.rb
@@ -18,7 +18,7 @@ module Avo
         field_id: field.id,
         via_record_id: via_record_id,
         via_resource_class: via_resource_class,
-        association_name: @field&.attribute_id || params[:related_name],
+        association_name: @field&.attribute_id || params[:association_name] || params[:related_name],
         encoded_filters: @current_encoded_filters,
         scope: params[:scope],
         q: params[:q]
